requirements.txt
yt-dlp>=2024.0.0
pillow>=10.0.0
cairosvg>=2.7.0
pyinstaller>=6.0.0
